pyproject.toml
[project]
name = "migratowl"
version = "0.1.0"
description = "AI-powered dependency migration analyzer"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2,<2",
    "langgraph>=1.0.7,<2",
    "langchain-core>=1.2,<2",
    "langchain-chroma>=1.1,<2",
    "langchain-openai>=1.1,<2",
    "tree-sitter>=0.25,<1",
    "tree-sitter-language-pack>=0.13",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "typer>=0.12",
    "rich>=13",
    "httpx>=0.27",
    "html2text>=2024.2",
    "packaging>=23",
    "fastmcp>=3.1,<4",
    "python-dotenv>=1",
]

[project.optional-dependencies]
anthropic = ["langchain-anthropic>=1,<2"]
google = ["langchain-google-vertexai>=3,<4", "langchain-google-genai>=2,<3"]
mistral = ["langchain-mistralai>=1,<2"]
groq = ["langchain-groq>=1,<2"]
all-providers = [
    "langchain-anthropic>=1,<2",
    "langchain-google-vertexai>=3,<4",
    "langchain-google-genai>=2,<3",
    "langchain-mistralai>=1,<2",
    "langchain-groq>=1,<2",
]

[project.scripts]
migratowl = "migratowl.interfaces.cli:app"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--ignore=tests/integration -W error -W 'ignore::pytest.PytestUnraisableExceptionWarning'"
markers = ["integration: integration tests requiring external services (network)"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["^\\.venv/", "^tests/integration/"]
